Exercise 3 TrolleyTotal

+++ WooliesX/DTO/Product.cs
@@ -2,7 +2,7 @@
 {
     public class Product
     {
-        public decimal Quantity { get; set; }
+        public long Quantity { get; set; }
         public string Name { get; set; }
         public decimal Price { get; set; }
     }
